Fixed bugs

+++ src/main/resources/db/migration/V3__Add_subscriptions_table.sql
@@ -1,5 +1,5 @@
 create table user_subscriptions (
-    channel_id int8 not null references users,
-    subscriber_id int8 not null references users,
+    channel_id bigint not null references users,
+    subscriber_id bigint not null references users,
     primary key (channel_id,subscriber_id)
 )
